Every page has 1 of the following arrangements of parent wrapper around the cointent that sits within .content__body wrapper we get from the HMRC template:
.grid-row > .column-one-third
.grid-row > .column-two-thirds
.column-two-thirds
.column-one-third
The majority of the time none of these classes are doing anything.
So at best we have 1 or 2 extar content div wrapper that are
redundant. But every now and then they are actually picking up the
styles for the classes assigned to the divs. Meaning that the
content displays as either 1/3 or 2/3 of the the parent
.content__body which is already set to 61.6666667% width, which
breaks the layout.