Uplift HMRC AF to v3 and uplift styles
A move to uplift to AF v3, which is a slow move towards eventually
moving away from AF all together.

As part of the work running through the app to see how AF v3 affects
the service layout I updated and fixed a couple of features like the
back button, spacing and font sizing around the continue button and
spacing on summary page continue button.